requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-decouple==3.8
pydantic==2.3.0
langchain==0.0.340
openai==0.27.6
tabulate==0.9.0
matplotlib==3.8.2
pandas==2.1.3
langchain-experimental==0.0.42
einops==0.7.0
accelerate==0.25.0
transformers==4.36.2